core: ask udev for modem capabilities
In combination with udev-extras, this allows automatic detection of modem
capabilties instead of having to keep 10-modem.fdi up-to-date with the hardware
flavor of the week.  NMHalManager grabs the originating device of a newly
detected device, passes that to NMManager, which then passes it along to the
device-type-specific creators.  The modem creator then asks udev for any
probed modem capabilities, and uses those in perference to any capabilities
HAL reports.  HAL capabilities are kept as a fallback.