modem: move probing into NetworkManager instead of udev-extras
For now.  It doesn't make a lot of sense to have modem probing in udev when
we switch to ModemManager in the future, since ModemManager would have to
re-do pretty much everything the udev-extras prober already did.  Thus,
since modem-probe from udev-extras doesn't have much of a future, don't
start using it or requiring it.  Move the code into NM instead, but still
use udev for the actual probing.