modem: remove originating device multiple serial port workaround
For now, handle multiple serial ports on a vendor/driver basis in the
modem creator function; HAL emits device added signals for the ttys in
random order, so just using the first serial port doesn't work.  This was
mainly added for 'hso' devices anyway, for which multiple serial ports are
already handled correctly by the creator.