modem: rework modem detection to better handle udev probing and multiple ports
Ignore multiple ports from the same modem device, becuase we can only use one
any exposing more than one in the UI is stupid and useless.  This requires some
drunken wandering through the HAL device tree to get the real device that all
the TTYs are owned by.

Second, prefer devices that both HAL and udev think are modems before using
a device that only udev thinks is a modem.  This requires that we delay
handling all ttys until HAL has finished sending them, so that if the second
tty is the one HAL thinks is a modem, we use that instead of some other port
that HAL sends first, but which might happen to have udev capabilities.