vpn: clear secrets when the connection fails
NM previously only cleared secrets when the VPN service daemon quit,
and the service daemons are on a 10-second inactivity timer.  So if
the user tried to re-activate the failed VPN connection within 10
seconds the old secrets would get used, which clearly isn't what we
want.  Ensure that whenever the VPN connection fails or disconnects,
we ask the settings service for secrets again the next time.