mbm: ignore driver when ignoring mbm ethernet devices
HAL makes it difficult to use driver matching for USB devices;
origdev_udi would point to the usb_interface object in HAL, which
always has a driver of 'usb'.  What we wanted to mbm was to just
look at the driver for the originating device, not the parent of
the originating device.  Whatever; there's no world where we want
mbm ethernet ports top-level devices anway, so the cdc_ether check
was superfluous anyway.