ipv6: fix incorrect address config signal emission
device->want_signal was never set to TRUE when addrconf was started,
causing random netlink events (say for link-local address addition
or removal) to trigger the config-changed signal from
nm_ip6_device_sync_from_netlink() at the wrong time.  This would
cause IPv6 address configuration to look like it succeeded, when
in fact the config timeout was still in-force.  Thus device
activation would proceed if IPv4 was enabled, but a few seconds later
the device would be deactivated due to the still active IPv6
timeout.

So fix that and clarify when the events from the IPv6 manager happen,
and what the want_signal variable is really for.