dnsmasq: VPN DNS fixes
Turn DHCP and DNS debugging on with NM_DNSMASQ_DEBUG.

Without --strict-order, dnsmasq will round-robin queries which in
the case of VPN connections may result in the query going to the
non-VPN nameserver.  Also, allow dnsmasq to poll resolv.conf for
nameserver updates so that when the default connection changes,
it knows about the new nameservers.