system-settings: cache secrets to protect against nm_connection_clear_secrets()
3rd patch in a series with:
0bbdc6b0fc3ca7425bb4e289571b014561b609d1
b135fa3265d1d684a93cb554a4c3f4690962f773

With NM 0.8 the system settings service was integrated into NM and
thus nm_connection_clear_secrets() acts directly on the system
settings plugins' NMConnection objects.  So when NM cleared secrets
(for example after determining that they might be bad in a device's
stage2 handler), we completely lost the secrets forever.

With this commit, the secrets are now cached and updated whenever
the connection is updated, and thus are again available to send to
NetworkManager when needed.