ifcfg-rh: fix distcheck and tests after 9e93924cca6c9f5f7fc6819babd9dd2771c1cc03
Since one test read in a file, wrote it out, and read it in again, we
have to be careful of whether srcdir == builddir or not.  If it doesn't,
then we need to remove the written-out connection file.  If it does, then
we don't want to remove that written-out connection file because it's
tracked by SCM.  Avoid the whole problem by writing it out to a separate
directory that we can always delete it from.