hal: fix driver name detection for USB wifi/eth devices
The previous code was using the originating device UDI as the
starting point, which was too far up the tree and so every
USB eth/wlan device's driver was showing up as 'usb' which was
pretty pointless.  Instead, just start at the actual device
and walk up the tree until we get a sensible driver name.