policy: fall back to original hostname before trying reverse DNS
Having the original hostname checks in the settings-service code
didn't allow the policy enough granularity to differentiate between
a plugin-provided hostname and the original hostname.  We want to
fall back to the original hostname if there isn't a persistent
hostname (from a plugin) and if there isn't a DHCP-provided
hostname.  Moving the original hostname checks to the policy
makes that possible.  Clarify the precedence order at the same
time, and minimally validate the DHCP hostname as well.