ifcfg-rh: provide an easy way to get connection UUID/path for tools
ifup/ifdown, nmcli, and possibly other tools need a way to get the
internally generated connection UUID and possibly it's path given
just the ifcfg file path.  Not all ifcfg files have the UUID in
the file, so ifcfg-rh generates one.  But external tools don't know
what that UUID is or what the logic that NM uses to generate it is.

Help them out a bit so that ifup/ifdown can easily poke NM given
just the ifcfg file.