wifi: prevent infinite loop when disposing of supplicant proxy (rh #538717)
This has been around a long time, but is very hard to trigger.  It appears
to happen mostly if the supplicant segfaults on resume but has been seen
in other cases as well.

For whatever reason, the DBusGProxy's refcount reaches 0 and the proxy gets
disposed of.  That in turn disposes of all the pending calls that are
in-progress on the proxy.  Since we give the pending calls a closure, that
closure (nm_supplicant_info_destroy) gets called when the pending calls
are destroyed.  That closure unrefs the proxy again.

Since DBusGProxy doesn't have any protection in its dispose() handler
against re-entrant disposes (which is arguably a bug of the client)
we end up infinite looping in nm_supplicant_info_destroy().

Fix that by ensuring we return early if we detect that we are already
freeing the NMSupplicantInfo object, and thus don't try to dispose
of the proxy yet again.