wifi: work around wl.o frequency reporting for active AP matching
Some proprietary drivers (wl.o) report tuned frequency (like when
scanning) instead of the associated AP's frequency.  This is a great
example of how WEXT is underspecified.  We use frequency to find the
active AP in the scan list because some configurations use the same
SSID/BSSID on the 2GHz and 5GHz bands simultaneously, and we need to
make sure we get the right AP in the right band.  This configuration
is uncommon though, and the frequency check penalizes closed drivers we
can't fix.  Because we're not total dicks, ignore the frequency condition
if the associated BSSID/SSID exists only in one band since that's most
likely the AP we want.